requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SocketIO==5.3.6
python-dotenv==1.0.0
Werkzeug==3.0.1
Pillow==10.1.0
PyMySQL==1.1.0
eventlet==0.34.3
requests==2.31.0